packages/vault/workers/trial-worker/requirements.txt
# Core
optuna[postgresql]>=3.6.0
psycopg2-binary>=2.9.9
redis>=5.0.0
sqlalchemy>=2.0.0

# ML/AI
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.4.0
rouge-score>=0.1.2
sentence-transformers>=2.5.0

# LLM APIs
openai>=1.30.0
anthropic>=0.28.0
httpx>=0.27.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.7.0
tenacity>=8.3.0
structlog>=24.1.0

# Monitoring
prometheus-client>=0.20.0
